refactor: Move base Builder up to PanicException
